Fix benchmarks
